+++ components/Chat/Chat.tsx
@@ -1,21 +1,19 @@
 import { Conversation, Message } from '@/types/chat';
-import { IconArrowDown } from '@tabler/icons-react';
 import { KeyValuePair } from '@/types/data';
 import { ErrorMessage } from '@/types/error';
 import { OpenAIModel, OpenAIModelID } from '@/types/openai';
 import { Prompt } from '@/types/prompt';
 import { throttle } from '@/utils';
-import { IconClearAll, IconKey, IconSettings } from '@tabler/icons-react';
-import { useTranslation } from 'next-i18next';
 import {
-  FC,
-  memo,
-  MutableRefObject,
-  useCallback,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
+  IconArrowDown,
+  IconClearAll,
+  IconKey,
+  IconPlayerStop,
+  IconRepeat,
+  IconSettings,
+} from '@tabler/icons-react';
+import { useTranslation } from 'next-i18next';
+import { FC, memo, MutableRefObject, useEffect, useRef, useState } from 'react';
 import { Spinner } from '../Global/Spinner';
 import { ChatInput } from './ChatInput';
 import { ChatLoader } from './ChatLoader';
@@ -25,38 +23,41 @@ import { ModelSelect } from './ModelSelect';
 import { SystemPrompt } from './SystemPrompt';
 
 interface Props {
-  conversation: Conversation;
-  models: OpenAIModel[];
   apiKey: string;
-  serverSideApiKeyIsSet: boolean;
+  conversation: Conversation;
+  conversationIsEmpty: boolean;
   defaultModelId: OpenAIModelID;
+  loading: boolean;
   messageIsStreaming: boolean;
   modelError: ErrorMessage | null;
-  loading: boolean;
-  prompts: Prompt[];
+  models: OpenAIModel[];
+  onEditMessage: (message: Message, messageIndex: number) => void;
+  onRegenerate: () => void;
   onSend: (message: Message, deleteCount?: number) => void;
   onUpdateConversation: (
     conversation: Conversation,
     data: KeyValuePair,
   ) => void;
-  onEditMessage: (message: Message, messageIndex: number) => void;
+  prompts: Prompt[];
+  serverSideApiKeyIsSet: boolean;
   stopConversationRef: MutableRefObject<boolean>;
 }
 
 export const Chat: FC<Props> = memo(
   ({
-    conversation,
-    models,
     apiKey,
-    serverSideApiKeyIsSet,
+    conversation,
+    conversationIsEmpty,
     defaultModelId,
+    loading,
     messageIsStreaming,
     modelError,
-    loading,
-    prompts,
+    models,
+    onEditMessage,
     onSend,
     onUpdateConversation,
-    onEditMessage,
+    prompts,
+    serverSideApiKeyIsSet,
     stopConversationRef,
   }) => {
     const { t } = useTranslation('chat');
@@ -69,13 +70,6 @@ export const Chat: FC<Props> = memo(
     const messagesEndRef = useRef<HTMLDivElement>(null);
     const chatContainerRef = useRef<HTMLDivElement>(null);
     const textareaRef = useRef<HTMLTextAreaElement>(null);
-
-    const scrollToBottom = useCallback(() => {
-      if (autoScrollEnabled) {
-        messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-        textareaRef.current?.focus();
-      }
-    }, [autoScrollEnabled]);
 
     const handleScroll = () => {
       if (chatContainerRef.current) {
@@ -102,6 +96,13 @@ export const Chat: FC<Props> = memo(
 
     const handleSettings = () => {
       setShowSettings(!showSettings);
+    };
+
+    const handleStopConversation = () => {
+      stopConversationRef.current = true;
+      setTimeout(() => {
+        stopConversationRef.current = false;
+      }, 1000);
     };
 
     const onClearAll = () => {
@@ -151,7 +152,7 @@ export const Chat: FC<Props> = memo(
     }, [messagesEndRef]);
 
     return (
-      <div className="overflow-hidden relative flex-1 bg-white dark:bg-[#343541]">
+      <div className="relative flex-1 overflow-hidden bg-white dark:bg-[#343541]">
         {!(apiKey || serverSideApiKeyIsSet) ? (
           <div className="mx-auto flex h-full w-[300px] flex-col justify-center space-y-6 sm:w-[500px]">
             <div className="mx-auto mb-5 text-gray-800 dark:text-gray-100">
@@ -285,35 +286,79 @@ export const Chat: FC<Props> = memo(
               )}
             </div>
 
-            <ChatInput
-              stopConversationRef={stopConversationRef}
-              textareaRef={textareaRef}
-              messageIsStreaming={messageIsStreaming}
-              conversationIsEmpty={conversation.messages.length === 0}
-              messages={conversation.messages}
-              model={conversation.model}
-              prompts={prompts}
-              onSend={(message) => {
-                setCurrentMessage(message);
-                onSend(message);
-              }}
-              onRegenerate={() => {
-                if (currentMessage) {
-                  onSend(currentMessage, 2);
-                }
-              }}
-            />
+            <div className="absolute bottom-0 left-0 w-full  border-transparent bg-gradient-to-b from-transparent via-white to-white px-4 pt-6 pb-2 dark:border-white/20 dark:via-[#343541] dark:to-[#343541] md:pt-2">
+              <div className="mx-auto flex flex-col gap-3 lg:max-w-3xl">
+                <div className="relative flex w-full flex-col items-center">
+                  {messageIsStreaming && (
+                    <button
+                      className="flex items-center gap-3 rounded border border-neutral-200 bg-white py-2 px-4 text-black hover:opacity-50 dark:border-neutral-600 dark:bg-[#343541] dark:text-white md:top-0"
+                      onClick={handleStopConversation}
+                    >
+                      <IconPlayerStop size={16} /> {t('Stop Generating')}
+                    </button>
+                  )}
+
+                  {!messageIsStreaming && !conversationIsEmpty && (
+                    <button
+                      className="flex items-center gap-3 rounded border border-neutral-200 bg-white py-2 px-4 text-black hover:opacity-50 dark:border-neutral-600 dark:bg-[#343541] dark:text-white md:top-0"
+                      onClick={() => {
+                        if (currentMessage) {
+                          onSend(currentMessage, 2);
+                        }
+                      }}
+                    >
+                      <IconRepeat size={18} /> {t('Regenerate response')}
+                    </button>
+                  )}
+
+                  {showScrollDownButton && (
+                    <div className="absolute bottom-0 right-0">
+                      <button
+                        className="flex h-7 w-7 items-center justify-center rounded-full bg-white shadow-md hover:shadow-lg dark:bg-[#515152d7]"
+                        onClick={handleScrollDown}
+                      >
+                        <IconArrowDown size={18} />
+                      </button>
+                    </div>
+                  )}
+                </div>
+
+                <ChatInput
+                  stopConversationRef={stopConversationRef}
+                  textareaRef={textareaRef}
+                  messageIsStreaming={messageIsStreaming}
+                  conversationIsEmpty={conversation.messages.length === 0}
+                  messages={conversation.messages}
+                  model={conversation.model}
+                  prompts={prompts}
+                  onSend={(message) => {
+                    setCurrentMessage(message);
+                    onSend(message);
+                  }}
+                  onRegenerate={() => {
+                    if (currentMessage) {
+                      onSend(currentMessage, 2);
+                    }
+                  }}
+                />
+
+                <div className="text-center text-[12px] text-black/50 dark:text-white/50">
+                  <a
+                    href="https://github.com/mckaywrigley/chatbot-ui"
+                    target="_blank"
+                    rel="noreferrer"
+                    className="underline"
+                  >
+                    ChatBot UI
+                  </a>
+                  .{' '}
+                  {t(
+                    "Chatbot UI is an advanced chatbot kit for OpenAI's chat models aiming to mimic ChatGPT's interface and functionality.",
+                  )}
+                </div>
+              </div>
+            </div>
           </>
-        )}
-        {showScrollDownButton && (
-          <div className="absolute bottom-0 right-0 mb-4 mr-4 pb-20">
-            <button
-              className="flex h-7 w-7 items-center justify-center rounded-full bg-white shadow-md hover:shadow-lg focus:outline-none focus:ring-2 focus:ring-blue-500 dark:bg-[#515152d7]"
-              onClick={handleScrollDown}
-            >
-              <IconArrowDown size={18} />
-            </button>
-          </div>
         )}
       </div>
     );
